requirements-dev.txt
boto3==1.26.115
click==8.1.3
faiss-cpu==1.7.3
numpy==1.24.2
openai==0.27.4
pandas==2.0.0
pytest==7.3.1
python-dotenv==1.0.0
scikit-learn==1.2.2
sentence-transformers==2.2.2
tokenizers==0.13.3
torch==2.0.0
tqdm==4.65.0
transformers==4.28.0
